@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.chad.baserecyclerviewadapterhelper"
         minSdkVersion 14
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 5
         versionName "2.1"
     }
@@ -28,9 +28,9 @@ dependencies {
     implementation project(path: ':library')
     implementation project(':material-spinner-1.0.5')
     implementation 'com.github.bumptech.glide:glide:3.7.0'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'com.google.android.material:material:1.0.0-rc02'
+    implementation 'androidx.cardview:cardview:1.0.0-rc02'
+    implementation 'androidx.appcompat:appcompat:1.0.0-rc02'
     implementation 'com.kyleduo.switchbutton:library:2.0.0'
     implementation 'com.orhanobut:logger:2.1.1'
 }
