@@ -1,8 +1,8 @@
 package com.chad.library.adapter.base;
 
-import android.support.annotation.IntRange;
-import android.support.annotation.LayoutRes;
-import android.util.Log;
+import androidx.annotation.IntRange;
+import androidx.annotation.LayoutRes;
+
 import android.util.SparseIntArray;
 import android.view.ViewGroup;
 
