@@ -1,7 +1,7 @@
 package com.chad.library.adapter.base;
 
-import android.support.annotation.IntRange;
-import android.support.annotation.LayoutRes;
+import androidx.annotation.IntRange;
+import androidx.annotation.LayoutRes;
 import android.util.SparseIntArray;
 import android.view.ViewGroup;
 
